requirements.txt
Django==1.10.1
Pillow==3.2.0
django-ajax-upload-widget==0.5.5
django-allauth==0.27.0
django-bootstrap-form==3.2.1
django-categories==1.4.2
 django-mptt-urls==2.0.3
